tests/wpt/web-platform-tests/page-visibility/unload-bubbles.html

<!doctype html>
<title>visibilitychange event bubbles when fired on unload</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function(t) {
  var w = window.open("resources/unload-bubbles.html");
  window.onmessage = t.step_func(function(event) {
    if (event.data === "close") {
      w.close();
      return;
    }

    const {state, target, bubbles} = event.data
    assert_equals(state, "hidden");
    assert_equals(target, "#document");
    assert_equals(bubbles, true);
    t.done();
  });
});
</script>
